[WIP] Implementing modular testing
