+++ meerkat/columns/image_column.py
@@ -2,19 +2,46 @@ from __future__ import annotations
 
 import logging
 from typing import Collection, Sequence
+from PIL.Image import Image
 
 import pandas as pd
 
 from meerkat.cells.imagepath import ImagePath
 from meerkat.columns.abstract import AbstractColumn
 from meerkat.columns.cell_column import CellColumn
-from meerkat.columns.lambda_column import LambdaColumn
+from meerkat.columns.lambda_column import LambdaColumn, LambdaCell
 from meerkat.columns.pandas_column import PandasSeriesColumn
 from meerkat.tools.lazy_loader import LazyLoader
 
 folder = LazyLoader("torchvision.datasets.folder")
 
 logger = logging.getLogger(__name__)
+
+
+class ImageCell(LambdaCell):
+    def __init__(
+        self,
+        transform: callable = None,
+        loader: callable = None,
+        data: str = None,
+    ):
+        self.loader = self.default_loader if loader is None else loader
+        self.transform = transform
+        self._data = data
+
+    def fn(self, filepath: str):
+        image = self.loader(filepath)
+        if self.transform is not None:
+            image = self.transform(image)
+        return image
+
+    def __eq__(self, other):
+        return (
+            (other.__class__ == self.__class__)
+            and (self.data == other.data)
+            and (self.transform == other.transform)
+            and (self.loader == other.loader)
+        )
 
 
 class ImageColumn(LambdaColumn):
@@ -26,11 +53,12 @@ class ImageColumn(LambdaColumn):
         *args,
         **kwargs,
     ):
-        super(ImageColumn, self).__init__(
-            PandasSeriesColumn.from_data(data), *args, **kwargs
-        )
+        super(ImageColumn, self).__init__(PandasSeriesColumn(data), *args, **kwargs)
         self.loader = self.default_loader if loader is None else loader
         self.transform = transform
+
+    def _create_cell(self, data: object) -> ImageCell:
+        return ImageCell(data=data, loader=self.loader, transform=self.transform)
 
     def fn(self, filepath: str):
         image = self.loader(filepath)
@@ -67,13 +95,12 @@ class ImageColumn(LambdaColumn):
         return "ImageCell(" + self.data.data.reset_index(drop=True) + ")"
 
     def is_equal(self, other: AbstractColumn) -> bool:
-        if other.__class__ != self.__class__:
-            return False
-        if self.loader != other.loader:
-            return False
-        if self.transform != other.transform:
-            return False
-        return self.data.is_equal(other.data)
+        return (
+            (other.__class__ == self.__class__)
+            and (self.loader == other.loader)
+            and (self.transform == other.transform)
+            and self.data.is_equal(other.data)
+        )
 
 
 class ImageCellColumn(CellColumn):
