+++ meerkat/mixins/inspect_fn.py
@@ -53,6 +53,8 @@ class FunctionInspectorMixin:
 
         # lazy import to avoid circular dependency
         from meerkat.columns.abstract import AbstractColumn
+        from meerkat.columns.tensor_column import TensorColumn
+        from meerkat.columns.numpy_column import NumpyArrayColumn
 
         if isinstance(output, Mapping):
             # `function` returns a dict output
@@ -75,8 +77,14 @@ class FunctionInspectorMixin:
 
         elif (
             isinstance(output, (bool, np.bool_))
-            or (isinstance(output, np.ndarray) and output.dtype == np.bool)
-            or (isinstance(output, torch.Tensor) and output.dtype == torch.bool)
+            or (
+                isinstance(output, (np.ndarray, NumpyArrayColumn))
+                and output.dtype == np.bool
+            )
+            or (
+                isinstance(output, (torch.Tensor, TensorColumn))
+                and output.dtype == torch.bool
+            )
         ):
 
             # `function` returns a bool
