+++ tests/meerkat/columns/test_image_column.py
@@ -1,7 +1,12 @@
 """Unittests for NumpyColumn."""
 from __future__ import annotations
+from meerkat.columns.image_column import ImageCell
+from meerkat.columns.pandas_column import PandasSeriesColumn
+from meerkat.columns.lambda_column import LambdaCell
 
 import os
+from typing import List, Union
+from _pytest.mark import param
 
 import numpy as np
 import pytest
@@ -13,6 +18,7 @@ from torchvision.transforms.functional import to_tensor
 from meerkat import ImageColumn
 from meerkat.columns.list_column import ListColumn
 from meerkat.columns.tensor_column import TensorColumn
+from meerkat.columns.abstract import AbstractColumn
 
 from .abstract import AbstractColumnTestBed, TestAbstractColumn
 
@@ -42,34 +48,141 @@ class ImageColumnTestBed(AbstractColumnTestBed):
             self.ims.append(im)
             self.data.append(transform(im) if transform else im)
             im.save(self.image_paths[-1])
-
+        if transform is not None:
+            self.data = torch.stack(self.data)
         self.transform = transform
         self.col = ImageColumn.from_filepaths(self.image_paths, transform=transform)
 
     def get_map_spec(
-        self, key: str = "map1", batched: bool = True, materialize: bool = False
+        self,
+        batched: bool = True,
+        materialize: bool = False,
+        salt: int = 1,
     ):
         if not materialize:
-            pass
+            if batched:
+                return {"fn": lambda x: x, "expected_result": self.col}
+            else:
+                # can't check for cell column equivalence because the `fn` is a bound
+                # method of different objects (since we perform batching then convert)
+                # non-batched fns to batched functions, so we call get
+                if self.transform is None:
+                    return {
+                        "fn": lambda x: x.get().rotate(45 + salt),
+                        "expected_result": ListColumn(
+                            [im.rotate(45 + salt) for im in self.ims]
+                        ),
+                    }
+                else:
+                    return {
+                        "fn": lambda x: x.get() + salt,
+                        "expected_result": TensorColumn(
+                            torch.stack([self.transform(im) for im in self.ims]) + salt
+                        ),
+                    }
         else:
             if self.transform is None:
                 return {
-                    "fn": lambda x: x.rotate(45),
-                    "expected_result": ListColumn([im.rotate(45) for im in self.ims]),
+                    "fn": (lambda x: [im.rotate(45 + salt) for im in x])
+                    if batched
+                    else (lambda x: x.rotate(45 + salt)),
+                    "expected_result": ListColumn(
+                        [im.rotate(45 + salt) for im in self.ims]
+                    ),
                 }
             else:
-                print(np.stack(self.image_arrays).shape)
                 return {
-                    "fn": lambda x: x + 1,
-                    "expected_result": TensorColumn(np.stack(self.image_arrays) + 1),
+                    "fn": lambda x: x + salt,
+                    "expected_result": TensorColumn(
+                        torch.stack([self.transform(im) for im in self.ims]) + salt
+                    ),
                 }
 
-    def get_data(self, index):
-        return self.data[index]
+    def get_filter_spec(
+        self,
+        batched: bool = True,
+        materialize: bool = False,
+        salt: int = 1,
+    ):
+        if not materialize:
+            if batched:
+                return {
+                    "fn": lambda x: [
+                        int(os.path.splitext(os.path.basename(cell.data))[0])
+                        < (4 + salt)
+                        for cell in x.lz
+                    ],
+                    "expected_result": self.col.lz[: 4 + salt],
+                }
+            else:
+                return {
+                    "fn": (
+                        lambda x: int(os.path.splitext(os.path.basename(x.data))[0])
+                        < (4 + salt)
+                    ),
+                    "expected_result": self.col.lz[: 4 + salt],
+                }
+        else:
+            if self.transform is None:
+                return {
+                    "fn": (lambda x: [im.rotate(45 + salt) for im in x])
+                    if batched
+                    else (lambda x: x.rotate(45 + salt)),
+                    "expected_result": ListColumn(
+                        [im.rotate(45 + salt) for im in self.ims]
+                    ),
+                }
+            else:
+                return {
+                    "fn": lambda x: (
+                        (x.mean(dim=[1, 2, 3]) if batched else x.mean()) > salt
+                    ).to(bool),
+                    "expected_result": self.col.lz[
+                        torch.stack([self.transform(im) for im in self.ims])
+                        .mean(dim=[1, 2, 3])
+                        .numpy()
+                        > salt
+                    ],
+                }
+
+    def get_data(self, index, materialize: bool = True):
+        if materialize:
+            if isinstance(index, int):
+                return self.data[index]
+
+            if self.transform is not None:
+                return self.data[index]
+            else:
+                index = np.arange(len(self.data))[index]
+                return [self.data[idx] for idx in index]
+        else:
+            if isinstance(index, int):
+                return ImageCell(
+                    data=self.image_paths[index],
+                    loader=self.col.loader,
+                    transform=self.col.transform,
+                )
+            index = np.arange(len(self.data))[index]
+            return PandasSeriesColumn([self.image_paths[idx] for idx in index])
 
     @staticmethod
-    def assert_data_equal(data1: np.ndarray, data2: np.ndarray):
-        assert (data1 == data2).all()
+    def assert_data_equal(
+        data1: Union[Image.Image, AbstractColumn, List, torch.Tensor],
+        data2: Union[Image.Image, AbstractColumn, List, torch.Tensor],
+    ):
+        if isinstance(data1, Image.Image) or isinstance(data1, List):
+            assert data1 == data2
+        elif isinstance(data1, AbstractColumn):
+            assert data1.is_equal(data2)
+        elif torch.is_tensor(data1):
+            assert (data1 == data2).all()
+        elif isinstance(data1, LambdaCell):
+            assert data1 == data2
+        else:
+            raise ValueError(
+                "Cannot assert data equal between objects type:"
+                f" {type(data1), type(data2)}"
+            )
 
 
 @pytest.fixture
@@ -87,28 +200,36 @@ class TestImageColumn(TestAbstractColumn):
     def _get_data_to_set(self, testbed, data_index):
         return np.zeros_like(testbed.get_data(data_index))
 
-    @ImageColumnTestBed.parametrize(params={"index_type": [np.array]})
+    @ImageColumnTestBed.parametrize(single=True, params={"index_type": [np.ndarray]})
     def test_set_item(self, testbed, index_type: type):
-        return super().test_set_item(testbed, index_type=index_type)
+        with pytest.raises(ValueError, match="Cannot setitem on a `LambdaColumn`."):
+            testbed.col[0] = 0
 
     @ImageColumnTestBed.parametrize(params={"index_type": [np.array]})
     def test_getitem(self, testbed, index_type: type):
         return super().test_getitem(testbed, index_type=index_type)
 
     @ImageColumnTestBed.parametrize(
-        config={"num_dims": [1], "dim_length": [1]}, params={"batched": [True, False]}
+        config={"transform": [True]},
+        params={"batched": [True, False], "materialize": [True, False]},
     )
-    def test_filter_1(self, testbed: AbstractColumnTestBed, batched: bool):
-        return super().test_filter_1(testbed, batched)
+    def test_filter_1(
+        self, testbed: AbstractColumnTestBed, batched: bool, materialize: bool
+    ):
+        return super().test_filter_1(testbed, batched, materialize=materialize)
 
-    @ImageColumnTestBed.parametrize(params={"batched": [True, False]})
+    @ImageColumnTestBed.parametrize(
+        params={"batched": [True, False], "materialize": [True, False]}
+    )
     def test_map_return_multiple(
         self, testbed: AbstractColumnTestBed, batched: bool, materialize: bool
     ):
-        return super().test_map_return_multiple(testbed, batched)
+        return super().test_map_return_multiple(
+            testbed, batched, materialize=materialize
+        )
 
     @ImageColumnTestBed.parametrize(
-        params={"batched": [True, False], "materialize": [True]}
+        params={"batched": [True, False], "materialize": [True, False]}
     )
     def test_map_return_single(
         self, testbed: AbstractColumnTestBed, batched: bool, materialize: bool
@@ -121,42 +242,20 @@ class TestImageColumn(TestAbstractColumn):
 
     @ImageColumnTestBed.parametrize()
     def test_io(self, tmp_path, testbed):
-        super().test_io(tmp_path, testbed)
+        # uses the tmp_path fixture which will provide a
+        # temporary directory unique to the test invocation,
+        # important for dataloader
+        col, _ = testbed.col, testbed.data
+
+        path = os.path.join(tmp_path, "test")
+        col.write(path)
+
+        new_col = self.column_class.read(path)
+
+        assert isinstance(new_col, self.column_class)
+        # can't check if the functions are the same since they point to different methods
+        assert col.data.is_equal(new_col.data)
 
     @ImageColumnTestBed.parametrize()
     def test_pickle(self, testbed):
         super().test_pickle(testbed)
-
-    @ImageColumnTestBed.parametrize()
-    def test_to_tensor(self, testbed):
-        col, _ = testbed.col, testbed.data
-
-        tensor = col.to_tensor()
-
-        assert torch.is_tensor(tensor)
-        assert (col == tensor.numpy()).all()
-
-    def test_from_array(self):
-        # Build a dataset from a batch
-        array = np.random.rand(10, 3, 3)
-        col = NumpyArrayColumn.from_array(array)
-
-        assert (col == array).all()
-        np_test.assert_equal(len(col), 10)
-
-    @ImageColumnTestBed.parametrize()
-    def test_to_pandas(self, testbed):
-        series = testbed.col.to_pandas()
-
-        assert isinstance(series, pd.Series)
-
-        if testbed.col.shape == 1:
-            assert (series.values == testbed.col.data).all()
-        else:
-            for idx in range(len(testbed.col)):
-                assert (series.iloc[idx] == testbed.col[idx]).all()
-
-    @ImageColumnTestBed.parametrize()
-    def test_repr_pandas(self, testbed):
-        series = testbed.col.to_pandas()
-        assert isinstance(series, pd.Series)
